Generalised typing for query operators
