+++ src/packages/pongo/src/core/collection/index.ts
@@ -1,0 +1,2 @@
+export * from './pongoCollection';
+export * from './query';
