+++ src/packages/pongo/src/core/index.ts
@@ -1,5 +1,5 @@
+export * from './collection';
 export * from './dbClient';
 export * from './pongoClient';
-export * from './pongoCollection';
 export * from './pongoSession';
 export * from './typing';
